pyproject.toml
[tool.black]
line-length = 100
target-version = ["py311"]
include = '\\.(py)$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
strict_equality = true
check_untyped_defs = true
pretty = true
plugins = []
exclude = [
    "venv",
    "build",
    "dist",
]

[build-system]
requires = ["setuptools>=62", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "credence"
version = "0.1.0"
description = "Credence — Pluggable Karma & Trust Engine"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Credence Maintainers"}]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "SQLAlchemy>=2.0.29",
    "PyYAML>=6.0.1",
    "python-jose[cryptography]>=3.3.0",
    "psycopg[binary]>=3.1.19",
    "alembic>=1.13.1",
    "redis>=5.0.4",
    "celery>=5.3.6",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "slowapi>=0.1.9",
    "requests>=2.31.0",
    "opentelemetry-instrumentation-fastapi>=0.47b0; python_version>='3.11'",
    "opentelemetry-instrumentation-sqlalchemy>=0.47b0; python_version>='3.11'",
]


